phpstan lvl 3

+++ classes/models/LengowCustomer.php
@@ -59,7 +59,7 @@ class LengowCustomer extends Customer
         $this->lastname = $data['last_name'];
         $this->fullName = $data['full_name'];
         $this->passwd = md5(rand());
-        $this->id_gender = LengowGender::getGender((string) $data['civility']);
+        $this->id_gender = (int) LengowGender::getGender((string) $data['civility']);
         $this->id_default_group = LengowConfiguration::get(LengowConfiguration::ORDER_CUSTOMER_GROUP);
 
         return $this;
